Refactor(be): `common` 모듈로 `exception`, `config` 이동
