--- backend/grabtable/src/main/java/edu/skku/grabtable/common/exception/BusinessException.java
@@ -1,10 +1,0 @@
-package edu.skku.grabtable.common.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class BusinessException extends RuntimeException {
-    private ErrorCode errorCode;
-}
