+++ backend/grabtable/src/main/java/edu/skku/grabtable/common/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package edu.skku.grabtable.exception;
+package edu.skku.grabtable.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
