+++ backend/grabtable/src/main/java/edu/skku/grabtable/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package edu.skku.grabtable.exception;
+package edu.skku.grabtable.common.exception;
 
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
