+++ backend/grabtable/src/test/java/edu/skku/grabtable/domain/UserTest.java
@@ -1,6 +1,6 @@
 package edu.skku.grabtable.domain;
 
-import edu.skku.grabtable.config.JpaAuditingConfig;
+import edu.skku.grabtable.common.config.JpaAuditingConfig;
 import edu.skku.grabtable.repository.UserRepository;
 import java.util.List;
 import org.assertj.core.api.Assertions;
